Moved config mgmt code to new package
